+++ lib/qtgui_time_plot_impl.cc
@@ -81,7 +81,7 @@ void qtgui_time_plot_impl::handle_msg(pmt::pmt_t msg)
     bool item_found_y;
     // Go through msg and search for key symbol
     item_found_y = false;
-    for (int k = 0; k < size_msg; k++) {
+    for (size_t k = 0; k < size_msg; k++) {
         msg_part = pmt::nth(k, msg);
         if (pmt::symbol_to_string(pmt::nth(0, msg_part)) == d_label_y.c_str()) {
             d_y = pmt::f32vector_elements(pmt::nth(1, msg_part));
